# Use `noalias()` in subset-assignment