--- a/R/yank.R
+++ b/R/yank.R
@@ -99,10 +99,10 @@ rray_yank_impl <- function(x, i) {
   i <- as_yank_indexer(i, x)
 
   # TODO
-  if (is.integer(i) && is_any_na_int(i)) {
+  if (is.integer(i) && is_any_na_int(list(i))) {
     abort("`NA` indices are not yet supported.")
   }
-  else if (is.logical(i) && is_any_na_int(as.integer(i))) {
+  else if (is.logical(i) && is_any_na_int(list(as.integer(i)))) {
     abort("`NA` indices are not yet supported.")
   }